Revert "initGL optimization changes (#2102)"
This reverts commit 5a6cb53beab9a51f267a9ed1410a6395e7c92b55.
