+++ examples/pxScene2d/src/gles2.js
@@ -24,8 +24,9 @@ var init = function(options) {
     var fullscreen = !!options.fullscreen;
     var title = options.title || "";
 
-    //gles2.init(width, height, fullscreen, title)
-    return global.sparkwebgl;
+    //gles2.init(width, height, fullscreen, title);
+
+    return global.sparkwebgl.instance;
 };
 
 module.exports = {
