+++ examples/pxScene2d/src/pxScene2d.cpp
@@ -155,7 +155,7 @@ extern uv_async_t asyncNewScript;
 extern uv_async_t gcTrigger;
 #endif // RUNINMAIN
 #endif //ENABLE_RT_NODE
-extern char *initAppScript;
+
 #ifdef ENABLE_VALGRIND
 #include <valgrind/callgrind.h>
 void startProfiling()
@@ -3263,9 +3263,8 @@ void pxScriptView::runScript()
       beginDrawing();
       glClearColor(0, 0, 0, 0);
       glClear(GL_COLOR_BUFFER_BIT);      
-      // compile initGL.js
-      mCtx->runScript(initAppScript);
-      rtValue foo = mCtx->get("loadAppUrl");
+      mCtx->runFile("initGL.js");
+      rtValue foo = mCtx->get("loadUrl");
       rtFunctionRef f = foo.toFunction();
       bool b = true;
       rtString url = mUrl;
